Add compilation of examples with new features to test suite
