--- examples/panic.rs
@@ -1,10 +1,0 @@
-// Triggers the panic handler. Should make all LEDs flash.
-
-#![no_std]
-
-use libtock::result::TockResult;
-
-#[libtock::main]
-async fn main() -> TockResult<()> {
-    panic!("Bye world!");
-}
